# New implementation of Env.make_copy_of_types