--- a/Changes
+++ b/Changes
@@ -30,6 +30,10 @@ Working version
   (Gabriel Scherer and Thomas Refis, review by Stephen Dolan and
    Florian Angeletti)
 
+- #8774: New implementation of Env.make_copy_of_types
+  (Alain Frisch, review by Thomas Refis, Leo White and Jacques Garrigue)
+
+
 ### Code generation and optimizations:
 
 - #8672: Optimise Switch code generation on booleans.