For reasons I cannot claim I fully understand, the type-checker copies the type of value declarations corresponding to possibly free variables in pattern matching branches.  This happens only in non-principal mode and for cases whose pattern possibly contains GADT constructors.

This current implementation computes an upper bound of free variables by traversing the parsetree.   This is one of the few ad hoc traversals of the parsetree (in addition to the main type-checking traversal), and I'd like to remove them as much as possible (secret goal: being able to register type-checking hooks to allowing e.g. expanding extension nodes knowing the actual typing environment).

So, this PR reimplements the copying of types in a different way.  See the code for details.  The type of an identifier is copied when that identifier is first accessed by the type-checker (instead of being copied before branches are type-checked).  I don't think that it should matter, but I'm not 100% sure.