@gasche 2019-06-28
Could we discuss the secret goal a bit more? It's not completely clear to me why typeful ppxes (ppt ?) requires to get rid of AST traversals. Is the idea that the extension points in question would be translated into a "temporary" parsetree that should type-check in a reasonable way (this is what eg. ReLit does), but you don't want those to be traversed by extra computations with various expectations? 

(I'm surprised by the removal of `iter_expression` that sounds like a useful helper function to have in the compiler, and possibly for compiler-libs users. Or is there a better alternative already?)