@trefis 2019-06-28
> For reasons I cannot claim I fully understand, the type-checker copies the type of value declarations corresponding to possibly free variables in pattern matching branches. This happens only in non-principal mode and for cases whose pattern possibly contains GADT constructors.

I had to look at that myself in the past, and since it isn't exactly obvious, nor is it documented, let me try to give a rough explanation of what is happening.

In the body of clauses which match on a GADT, we might introduce [ambivalence](http://gallium.inria.fr/~remy/gadts/Garrigue-Remy:gadts@aplas2013.pdf), but we want to restrict the scope of such ambivalent types to the branch introducing the equation.
Take this example:
```ocaml
type (_,_) eq = Refl : ('a,'a) eq

let cast : type a b. (a, b) eq -> a -> b =
  fun eq x ->
    match eq with Refl -> x
```
When we arrive in the body of the branch, we know that `x` has type `a`, and we need to modify that type to make it `a ≈ b`.

However, and this is actually an implementation detail, what the typechecker actually knows in the body of the branch is that `x` is an instance of type `a` at a [level](http://okmij.org/ftp/ML/generalization.html) `l`, which is outside of the scope of the equation. And that actually prevents us from introducing the ambivalence we wanted (because it would be leaking outside of the equation's scope).

The copy of the type of `x` solves that problem: it distinguishes the type used outside the body of the branch from the one used inside (i.e. no leaking), and in doing so updates the level of the one used inside the branch.

This is not needed in principal mode, because in that mode we're much more careful about the amount of sharing that happens between types, which things are known principally and which aren't, etc.

---

As for the actual implementation, I'll review it once there is a high-level agreement (with @gasche, and possibly others) about the acceptance of the change.