@alainfrisch 2019-06-28
The secret goal could be to allow hooking a custom `type_expect` function (for any expression, or restricted to extension nodes, to be seen).  Depending on the version of the secret goal, the custom function could rewritte the Parsetree fragment, or directly do the type-checking, producing Typedtree nodes.  In both cases, the function could access the current type-checking environment (otherwise it would be a plain ppx).  Each extra traversal, in addition to type_expect, would need to be overridden in the same way, but sometimes this is not possible (if the "expansion" depends on the type-checking environment, or the expected type).

> (I'm surprised by the removal of iter_expression that sounds like a useful helper function to have in the compiler, and possibly for compiler-libs users. Or is there a better alternative already?)

Ast_iterator is more powerful.