@ksromanov 2019-06-29
> secret goal: being able to register type-checking hooks to allowing e.g. expanding extension nodes knowing the actual typing environment.

Does it mean that if everything goes well, in a year or two it will be possible to use type information in PPX plugins? Or, at least make plugins to the Compiler, which have access both to initial Parsetree and the type info?