@alainfrisch 2019-07-01
I suggest to keep the discussion focused on this PR, which, I think, has merits independently of the "secret goal".