@alainfrisch 2019-07-03
I've pushed a commit that prevents repeated copying: after a lookup, only the outermost copying is applied.  With this change, the code produced by the script below compiles much faster (tested with ocamlc.opt): 1.35s against 6.6s on trunk (type-checking only: 0.6s against 5.8s).

````ocaml
let n1 = 2500
let n2 = 300
let n3 = 5

let () =
  let f () =
    for j = 1 to n2 do
      Printf.printf "method f%i = ()\n" j
    done
  in
  print_endline "type t = A : t";
  for i = 1 to n3 do
    Printf.printf "let a%i = object\n" i;
    f ();
    Printf.printf "end\n"
  done;
  print_endline "let f x =\n";
  for i = 1 to n1 do
    print_endline "match x with A -> ";
  done;
  for i = 1 to n3 do
    Printf.printf "a%i," i;
  done;
  print_endline "()"
````

Running the compiler with OCAMLRUNPARAM=v=1024:

trunk:
````
allocated_words: 568881603
minor_words: 568829267
promoted_words: 129001886
major_words: 129054222
minor_collections: 2178
major_collections: 17
heap_words: 132075520
heap_chunks: 41
top_heap_words: 132075520
compactions: 0
````

PR:
````
allocated_words: 278927663
minor_words: 278875324
promoted_words: 11814203
major_words: 11866542
minor_collections: 1073
major_collections: 19
heap_words: 3032064
heap_chunks: 14
top_heap_words: 3032064
compactions: 0
````